test: 메모리 저장 VoucherRepository 테스트
