+++ src/main/java/com/devcourse/springbootbasic/application/domain/voucher/dto/VoucherDto.java
@@ -1,0 +1,7 @@
+package com.devcourse.springbootbasic.application.domain.voucher.dto;
+
+public record VoucherDto(
+        VoucherType voucherType,
+        double discountValue
+) {
+}
