+++ src/main/java/com/devcourse/springbootbasic/application/domain/voucher/factory/VoucherFactory.java
@@ -1,0 +1,7 @@
+package com.devcourse.springbootbasic.application.domain.voucher.factory;
+
+import com.devcourse.springbootbasic.application.domain.voucher.data.Voucher;
+
+public interface VoucherFactory {
+    Voucher create(double voucherDiscount);
+}
