--- src/main/java/com/devcourse/springbootbasic/application/exception/InvalidDataException.java
@@ -1,16 +1,0 @@
-package com.devcourse.springbootbasic.application.exception;
-
-public class InvalidDataException extends RuntimeException {
-
-    public InvalidDataException() {
-        super();
-    }
-
-    public InvalidDataException(String errorMessage) {
-        super(errorMessage);
-    }
-
-    public InvalidDataException(String errorMessage, Throwable cause) {
-        super(errorMessage, cause);
-    }
-}
