--- src/main/java/com/devcourse/springbootbasic/application/factory/VoucherFactory.java
@@ -1,7 +1,0 @@
-package com.devcourse.springbootbasic.application.factory;
-
-import com.devcourse.springbootbasic.application.domain.Voucher;
-
-public interface VoucherFactory {
-    Voucher create(double voucherDiscount);
-}
