+++ src/main/java/com/devcourse/springbootbasic/application/global/exception/InvalidDataException.java
@@ -1,0 +1,16 @@
+package com.devcourse.springbootbasic.application.global.exception;
+
+public class InvalidDataException extends RuntimeException {
+
+    public InvalidDataException() {
+        super();
+    }
+
+    public InvalidDataException(String errorMessage) {
+        super(errorMessage);
+    }
+
+    public InvalidDataException(String errorMessage, Throwable cause) {
+        super(errorMessage, cause);
+    }
+}
