+++ src/test/java/com/devcourse/springbootbasic/application/dto/ListMenuTest.java
@@ -1,6 +1,7 @@
 package com.devcourse.springbootbasic.application.dto;
 
-import com.devcourse.springbootbasic.application.exception.InvalidDataException;
+import com.devcourse.springbootbasic.application.global.model.ListMenu;
+import com.devcourse.springbootbasic.application.global.exception.InvalidDataException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
